+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 * Fix escaping of interpolations after dollar signs.
 * Fix nixfmt trying to allocate temp files that aren't used.
+* Nixfmt now accepts the '-' argument to read from stdin.
+* `nixfmt [dir]` now recursively formats nix files in that directory.
 
 ## 0.5.0 -- 2022-03-15